+++ ninja-servlet-jpa-blog-archetype/src/main/resources/archetype-resources/src/main/java/conf/StartupActions.java
@@ -1,0 +1,39 @@
+#set( $symbol_pound = '#' )
+#set( $symbol_dollar = '$' )
+#set( $symbol_escape = '\' )
+package conf;
+
+import javax.inject.Singleton;
+
+import ninja.lifecycle.Start;
+import ninja.utils.NinjaProperties;
+
+import com.google.inject.Inject;
+
+import dao.SetupDao;
+
+@Singleton
+public class StartupActions {
+    
+    @Inject
+    SetupDao setupDao;
+
+    private NinjaProperties ninjaProperties;
+
+    @Inject
+    public StartupActions(NinjaProperties ninjaProperties) {
+        this.ninjaProperties = ninjaProperties;
+    }
+    
+    @Start(order=100)
+    public void generateDummyDataWhenInTest() {
+        
+        if (!ninjaProperties.isProd()) {
+            
+            setupDao.setup();
+            
+        }
+        
+    }
+
+}