+++ ninja-servlet-jpa-blog-archetype/src/main/resources/archetype-resources/src/main/java/models/ArticlesDto.java
@@ -1,0 +1,14 @@
+#set( $symbol_pound = '#' )
+#set( $symbol_dollar = '$' )
+#set( $symbol_escape = '\' )
+package models;
+
+import java.util.List;
+
+import models.Article;
+
+public class ArticlesDto {
+
+    public List<Article> articles;
+
+}